feat(dino): Move moving-input-process functionality
To move the dino, the pg event system us accesed to to save get the user inputs.
It has to be ensured that the dino state is changed only in specific cases.
Until now this check is done directly in the main file.
Now this behaviour is stored in the class directly.