tests(): Add GitHub Actions workflow to run tests
This commit adds a GitHub Actions workflow to run tests on every push to the main branch or pull request in general. The workflow is defined in the `.github/workflows/run_tests.yml` file.
It runs the tests on lates version of ubuntu and four different versions of python.
Thes versions run from 3.8 to 3.11.