refactor(): Change methods to privat if possible
Everywhere where it is possible, I changed methods to private. This is a good practice to hide the implementation details of a class. This prevents issues in the future by accidentally calling a method that should not be called from outside the class.
Unfortunatelly, this leads to confusion in the test modules. Based on the name mangeling, the private methods are not directly accessible from the test modules. This is why I had to change the test modules as well. I had to change the called method to the format `[instance]._[clas name]__[method name]` This is not a good practice that leads to a lot confusion because of missing type hints, but it is the only way to test them separately.